Revert "[VIR-208] 디스커션 생성 알림 기능 "
