+++ client/discussion/discussion_client.go
@@ -4,27 +4,241 @@ import (
 	"encoding/json"
 	"fmt"
 	"strconv"
+	"time"
+
+	"code.gitea.io/gitea/services/context"
 
 	"code.gitea.io/gitea/client"
-	"code.gitea.io/gitea/client/discussion/model"
+	repo_model "code.gitea.io/gitea/models/repo"
+	user_model "code.gitea.io/gitea/models/user"
 	"code.gitea.io/gitea/modules/log"
+	"code.gitea.io/gitea/modules/timeutil"
 	"github.com/go-resty/resty/v2"
 )
 
+type DiscussionCode struct {
+	Id        int64  `json:"id"`
+	FilePath  string `json:"filePath"`
+	StartLine int    `json:"startLine"`
+	EndLine   int    `json:"endLine"`
+}
+
+type DeleteDiscussionCommentRequest struct {
+	PosterId            int64 `json:"posterId"`
+	DiscussionCommentId int64 `json:"discussionCommentId"`
+}
+
+type PostDiscussionRequest struct {
+	RepoId     int64            `json:"repoId"`
+	Poster     *user_model.User `json:"-"`
+	PosterId   int64            `json:"posterId"`
+	Name       string           `json:"name"`
+	Content    string           `json:"content"`
+	BranchName string           `json:"branchName"`
+	Codes      []DiscussionCode `json:"codes"`
+}
+
+type DiscussionAvailableRequest struct {
+	RepoId    int64 `json:"repoId"`
+	Available bool  `json:"available"`
+}
+
+type CommentScopeEnum int
+
+const (
+	CommentScopeGlobal CommentScopeEnum = iota
+	CommentScopeLocal
+)
+
+type PostCommentRequest struct {
+	DiscussionId int64            `json:"discussionId"`
+	CodeId       *int64           `json:"codeId"`
+	GroupId      *int64           `json:"groupId"`
+	PosterId     int64            `json:"posterId"`
+	Scope        CommentScopeEnum `json:"scope"`
+	StartLine    *int32           `json:"startLine"`
+	EndLine      *int32           `json:"endLine"`
+	Content      string           `json:"content"`
+}
+
+type ModifyDiscussionRequest struct {
+	RepoId       int64            `json:"repoId"`
+	DiscussionId int64            `json:"discussionId"`
+	PosterId     int64            `json:"posterId"`
+	Name         string           `json:"name"`
+	Content      string           `json:"content"`
+	Codes        []DiscussionCode `json:"codes"`
+}
+
+type DiscussionResponse struct {
+	Id          int64              `json:"id"`
+	Name        string             `json:"name"`
+	Content     string             `json:"content"`
+	RepoId      int64              `json:"repoId"`
+	PosterId    int64              `json:"posterId"`
+	CommitHash  string             `json:"commitHash"`
+	IsClosed    bool               `json:"isClosed"`
+	Deadline    timeutil.TimeStamp `json:"deadline"`
+	Assignees   []int64            `json:"assignees"`
+	CreatedUnix timeutil.TimeStamp `json:"createdUnix"`
+	UpdatedUnix timeutil.TimeStamp `json:"updatedUnix"`
+	Index       int64              `json:"index"`
+	Poster      *user_model.User   `json:"-"`
+}
+
+type ReactionList []*DiscussionReaction
+
+func (list ReactionList) GroupByType() map[string]ReactionList {
+	reactions := make(map[string]ReactionList)
+	for _, reaction := range list {
+		reactions[reaction.Type] = append(reactions[reaction.Type], reaction)
+	}
+	return reactions
+}
+
+func (list ReactionList) HasUser(userId int64) bool {
+	if userId == 0 {
+		return false
+	}
+	for _, reaction := range list {
+		if reaction.UserId == userId {
+			return true
+		}
+	}
+	return false
+}
+
+type Discussion struct {
+	Id           int64                  `json:"id"`
+	Name         string                 `json:"name"`
+	Content      string                 `json:"content"`
+	RepoId       int64                  `json:"repoId"`
+	PosterId     int64                  `json:"posterId"`
+	CommitHash   string                 `json:"commitHash"`
+	Index        int64                  `json:"index"`
+	IsClosed     bool                   `json:"isClosed"`
+	CreatedUnix  timeutil.TimeStamp     `json:"createdUnix"` // required, but didn't exist before
+	ClosedUnix   timeutil.TimeStamp     `json:"closedUnix"`  // required, but didn't exist before
+	DeadlineUnix timeutil.TimeStamp     `json:"deadlineUnix"`
+	NumComments  int                    `json:"-"` // it can be computed
+	Repo         *repo_model.Repository `json:"-"` // it can be computed via d.LoadRepo()
+	Poster       *user_model.User       `json:"-"` // it can be computed via d.LoadPoster()
+}
+
+type DiscussionListResponse struct {
+	TotalCount  int64         `json:"totalCount"`
+	Discussions []*Discussion `json:"discussions"`
+}
+
+type ModifyDiscussionCommentRequest struct {
+	DiscussionId        int64            `json:"discussionId"`
+	DiscussionCommentId int64            `json:"discussionCommentId"`
+	CodeId              *int64           `json:"codeId"`
+	PosterId            int64            `json:"posterId"`
+	Scope               CommentScopeEnum `json:"scope"`
+	StartLine           *int32           `json:"startLine"`
+	EndLine             *int32           `json:"endLine"`
+	Content             string           `json:"content"`
+}
+
+type DiscussionCountResponse struct {
+	OpenCount  int `json:"openCount"`
+	CloseCount int `json:"closeCount"`
+}
+
+type ExtractedLine struct {
+	LineNumber int    `json:"lineNumber"`
+	Content    string `json:"content"`
+}
+type CodeBlock struct {
+	CodeId   int64                       `json:"codeId"`
+	Lines    []ExtractedLine             `json:"lines"`
+	Comments []DiscussionCommentResponse `json:"comments"`
+}
+type FileContent struct {
+	FilePath   string      `json:"filePath"`
+	CodeBlocks []CodeBlock `json:"codeBlocks"`
+}
+
+type DiscussionCommentResponse struct {
+	Id           int64                 `json:"id"`
+	FilePath     string                `json:"filePath"`
+	GroupId      int64                 `json:"groupId"`
+	DiscussionId int64                 `json:"discussionId"`
+	PosterId     int64                 `json:"posterId"`
+	CodeId       int64                 `json:"codeId"`
+	Scope        string                `json:"scope"`
+	StartLine    int64                 `json:"startLine"`
+	EndLine      int64                 `json:"endLine"`
+	Content      string                `json:"content"`
+	CreatedUnix  timeutil.TimeStamp    `json:"createdUnix"`
+	Reactions    []*DiscussionReaction `json:"reactions"`
+}
+
+type ReactionTypeEnum = string
+
+const (
+	PLUS_ONE  ReactionTypeEnum = "+1"
+	MINUS_ONE ReactionTypeEnum = "-1"
+	LAUGH     ReactionTypeEnum = "laugh"
+	HOORAY    ReactionTypeEnum = "hooray"
+	CONFUSED  ReactionTypeEnum = "confused"
+	HEART     ReactionTypeEnum = "heart"
+	ROCKET    ReactionTypeEnum = "rocket"
+	EYES      ReactionTypeEnum = "eyes"
+)
+
+type DiscussionReaction struct {
+	Id           int64            `json:"id"`
+	Type         ReactionTypeEnum `json:"type"`
+	DiscussionId int64            `json:"discussionId"`
+	CommentId    int64            `json:"commentId"`
+	UserId       int64            `json:"userId"`
+}
+type DiscussionContentResponse struct {
+	DiscussionId    int64                       `json:"discussionId"`
+	Contents        []FileContent               `json:"contents"`
+	GlobalComments  []DiscussionCommentResponse `json:"globalComments"`
+	GlobalReactions []DiscussionReaction        `json:"discussionReaction"`
+}
+
+type DiscussionDeadline struct {
+	Deadline *time.Time `json:"due_date"`
+}
+
+type UpdateAssigneeRequest struct {
+	DiscussionId int64 `json:"discussionId"`
+	AssigneeId   int64 `json:"assigneeId"`
+}
+
+type DiscussionErrorResponse struct {
+	TimeStamp timeutil.TimeStamp `json:"timestamp"`
+	Status    int                `json:"status"`
+	Error     string             `json:"error"`
+	message   string
+}
+
 func validateResponse(resp *resty.Response) error {
 	if resp.IsError() {
-		var errResp model.DiscussionErrorResponse
+		var errResp DiscussionErrorResponse
 		if err := json.Unmarshal(resp.Body(), &errResp); err != nil {
 			log.Error("Failed to parse error response: %v", err)
 			return fmt.Errorf("unexpected error: %s", resp.Status())
 		}
-		log.Error("API Error %d: %s", errResp.Status, errResp.Message)
-		return fmt.Errorf("api error %d: %s", errResp.Status, errResp.Message)
+		log.Error("API Error %d: %s", errResp.Status, errResp.message)
+		return fmt.Errorf("api error %d: %s", errResp.Status, errResp.message)
 	}
 	return nil
 }
 
-func PostDiscussion(request *model.PostDiscussionRequest) (int, error) {
+type DiscussionReactionRequest struct {
+	Type         ReactionTypeEnum `json:"type"`
+	DiscussionId int64            `json:"discussionId"`
+	CommentId    int64            `json:"commentId"`
+	UserId       int64            `json:"userId"`
+}
+
+func PostDiscussion(request *PostDiscussionRequest) (int, error) {
 	resp, err := client.Request().
 		SetBody(request).
 		Post("/discussion")
@@ -44,7 +258,7 @@ func PostDiscussion(request *model.PostDiscussionRequest) (int, error) {
 	return result, nil
 }
 
-func GetDiscussion(repoId int64) (*model.DiscussionResponse, error) {
+func GetDiscussion(repoId int64) (*DiscussionResponse, error) {
 	resp, err := client.Request().
 		Get(fmt.Sprintf("/discussion/%d", repoId))
 	if err != nil {
@@ -55,7 +269,7 @@ func GetDiscussion(repoId int64) (*model.DiscussionResponse, error) {
 		return nil, err
 	}
 
-	var result model.DiscussionResponse
+	var result DiscussionResponse
 	if err := json.Unmarshal(resp.Body(), &result); err != nil {
 		return nil, fmt.Errorf("failed to parse response body: %w", err)
 	}
@@ -63,7 +277,7 @@ func GetDiscussion(repoId int64) (*model.DiscussionResponse, error) {
 	return &result, nil
 }
 
-func GetDiscussionCount(repoId int64) (*model.DiscussionCountResponse, error) {
+func GetDiscussionCount(repoId int64) (*DiscussionCountResponse, error) {
 	resp, err := client.Request().
 		Get(fmt.Sprintf("/discussion/%d/count", repoId))
 	if err != nil {
@@ -74,7 +288,7 @@ func GetDiscussionCount(repoId int64) (*model.DiscussionCountResponse, error) {
 		return nil, err
 	}
 
-	var result model.DiscussionCountResponse
+	var result DiscussionCountResponse
 	if err := json.Unmarshal(resp.Body(), &result); err != nil {
 		return nil, fmt.Errorf("failed to parse response body: %w", err)
 	}
@@ -82,7 +296,7 @@ func GetDiscussionCount(repoId int64) (*model.DiscussionCountResponse, error) {
 	return &result, nil
 }
 
-func GetDiscussionList(repoId int64, isClosed bool, page int, sort string) (*model.DiscussionListResponse, error) {
+func GetDiscussionList(repoId int64, isClosed bool, page int, sort string) (*DiscussionListResponse, error) {
 	isClosedAsString := strconv.FormatBool(isClosed)
 	pageAsString := strconv.Itoa(page)
 	resp, err := client.Request().
@@ -98,7 +312,7 @@ func GetDiscussionList(repoId int64, isClosed bool, page int, sort string) (*mod
 		return nil, err
 	}
 
-	var result model.DiscussionListResponse
+	var result DiscussionListResponse
 	if err := json.Unmarshal(resp.Body(), &result); err != nil {
 		return nil, fmt.Errorf("failed to parse response body: %w", err)
 	}
@@ -106,7 +320,7 @@ func GetDiscussionList(repoId int64, isClosed bool, page int, sort string) (*mod
 	return &result, nil
 }
 
-func GetDiscussionContent(discussionId int64) (*model.DiscussionContentResponse, error) {
+func GetDiscussionContent(discussionId int64) (*DiscussionContentResponse, error) {
 	resp, err := client.Request().
 		Get(fmt.Sprintf("/discussion/%d/contents", discussionId))
 	if err != nil {
@@ -117,7 +331,7 @@ func GetDiscussionContent(discussionId int64) (*model.DiscussionContentResponse,
 		return nil, err
 	}
 
-	var result model.DiscussionContentResponse
+	var result DiscussionContentResponse
 	if err := json.Unmarshal(resp.Body(), &result); err != nil {
 		return nil, fmt.Errorf("failed to parse response body: %w", err)
 	}
@@ -125,7 +339,7 @@ func GetDiscussionContent(discussionId int64) (*model.DiscussionContentResponse,
 	return &result, nil
 }
 
-func GetDiscussionComment(discussionCommentId int64) (*model.DiscussionCommentResponse, error) {
+func GetDiscussionComment(discussionCommentId int64) (*DiscussionCommentResponse, error) {
 
 	resp, err := client.Request().
 		SetQueryParam("id", strconv.FormatInt(discussionCommentId, 10)).
@@ -139,7 +353,7 @@ func GetDiscussionComment(discussionCommentId int64) (*model.DiscussionCommentRe
 		return nil, err
 	}
 
-	var result model.DiscussionCommentResponse
+	var result DiscussionCommentResponse
 	if err := json.Unmarshal(resp.Body(), &result); err != nil {
 		return nil, fmt.Errorf("failed to parse response body: %w", err)
 	}
@@ -147,10 +361,10 @@ func GetDiscussionComment(discussionCommentId int64) (*model.DiscussionCommentRe
 	return &result, nil
 }
 
-func GetDiscussionCommentsByCodeId(codeId int64) ([]*model.DiscussionCommentResponse, error) {
+func GetDiscussionCommentsByCodeId(codeId int64) ([]*DiscussionCommentResponse, error) {
 	resp, err := client.Request().Get(fmt.Sprintf("/discussion/comments/%d", codeId))
 
-	result := make([]*model.DiscussionCommentResponse, 0)
+	result := make([]*DiscussionCommentResponse, 0)
 
 	if err = json.Unmarshal(resp.Body(), &result); err != nil {
 		return nil, fmt.Errorf("failed to unmarshal response: %w", err)
@@ -174,7 +388,7 @@ func HandleDiscussionAvailable() (*resty.Response, error) {
 	return resp, nil
 }
 
-func PostComment(request *model.PostCommentRequest) (*int64, error) {
+func PostComment(request *PostCommentRequest) (*int64, error) {
 	resp, err := client.Request().
 		SetBody(request).
 		Post("/discussion/comment")
@@ -195,7 +409,7 @@ func PostComment(request *model.PostCommentRequest) (*int64, error) {
 	return &id, nil
 }
 
-func ModifyDiscussion(request *model.ModifyDiscussionRequest) (*resty.Response, error) {
+func ModifyDiscussion(request *ModifyDiscussionRequest) (*resty.Response, error) {
 	resp, err := client.Request().
 		SetBody(request).
 		Put("/discussion")
@@ -211,7 +425,7 @@ func ModifyDiscussion(request *model.ModifyDiscussionRequest) (*resty.Response, 
 }
 
 func DeleteDiscussionComment(discussionCommentId int64, posterId int64) error {
-	request := &model.DeleteDiscussionCommentRequest{
+	request := &DeleteDiscussionCommentRequest{
 		DiscussionCommentId: discussionCommentId,
 		PosterId:            posterId,
 	}
@@ -230,12 +444,61 @@ func DeleteDiscussionComment(discussionCommentId int64, posterId int64) error {
 	return nil
 }
 
-func GetDiscussionContents(discussionId int64) (*model.DiscussionContentResponse, error) {
+/**
+ * discussion methods
+ * the `discussion` struct could be moved to a separate file later
+ */
+func (d *Discussion) IsExpired() bool {
+	return d.DeadlineUnix < timeutil.TimeStamp(time.Now().Unix())
+}
+
+func (d *Discussion) GetLastEventTimestamp() timeutil.TimeStamp {
+	if d.IsClosed {
+		return d.ClosedUnix
+	}
+	return d.CreatedUnix
+}
+
+func (d *Discussion) GetLastEventLabel() string {
+	if d.IsClosed {
+		return "repo.discussion.closed_by"
+	}
+	return "repo.discussion.opened_by"
+}
+
+func (d *Discussion) GetLastEventLabelFake() string {
+	if d.IsClosed {
+		return "repo.discussion.closed_by_fake"
+	}
+	return "repo.discussion.opened_by_fake"
+}
+
+func (d *Discussion) LoadPoster(ctx *context.Context) (err error) {
+	if d.Poster == nil && d.PosterId != 0 {
+		d.Poster, err = user_model.GetPossibleUserByID(ctx, d.PosterId)
+		if err != nil {
+			d.PosterId = user_model.GhostUserID
+			d.Poster = user_model.NewGhostUser()
+			if !user_model.IsErrUserNotExist(err) {
+				return fmt.Errorf("getUserById.(poster) [%d]: %w", d.PosterId, err)
+			}
+			return nil
+		}
+	}
+	return err
+}
+
+func (d *Discussion) LoadRepo(ctx *context.Context) (err error) {
+	d.Repo = ctx.Repo.Repository
+	return nil
+}
+
+func GetDiscussionContents(discussionId int64) (*DiscussionContentResponse, error) {
 	resp, err := client.Request().Get(fmt.Sprintf("/discussion/%d/contents", discussionId))
 	if err != nil {
 		return nil, err
 	}
-	result := &model.DiscussionContentResponse{}
+	result := &DiscussionContentResponse{}
 
 	if err := json.Unmarshal(resp.Body(), result); err != nil {
 		return nil, err
@@ -274,7 +537,11 @@ func SetDiscussionDeadline(discussionId int64, deadline int64) error {
 	return nil
 }
 
-func UpdateDiscussionAssignee(request *model.UpdateAssigneeRequest) error {
+func (dr DiscussionResponse) IsPoster(id int64) bool {
+	return dr.PosterId == id
+}
+
+func UpdateDiscussionAssignee(request *UpdateAssigneeRequest) error {
 	resp, err := client.Request().SetBody(request).Put("/discussion/assignees")
 	if err != nil {
 		return err
@@ -285,7 +552,7 @@ func UpdateDiscussionAssignee(request *model.UpdateAssigneeRequest) error {
 	return nil
 }
 
-func ModifyDiscussionComment(request *model.ModifyDiscussionCommentRequest) error {
+func ModifyDiscussionComment(request *ModifyDiscussionCommentRequest) error {
 	resp, err := client.Request().SetBody(request).Put("/discussion/comment")
 	if err != nil {
 		return err
@@ -309,7 +576,7 @@ func ClearDiscussionAssignee(discussionId int64) error {
 	return nil
 }
 
-func GiveReaction(request model.DiscussionReactionRequest) (int64, error) {
+func GiveReaction(request DiscussionReactionRequest) (int64, error) {
 	var result int64 = -1
 	resp, err := client.Request().SetBody(request).Post("/discussion/reaction")
 	if err != nil {
@@ -324,7 +591,7 @@ func GiveReaction(request model.DiscussionReactionRequest) (int64, error) {
 	return result, err
 }
 
-func RemoveReaction(request model.DiscussionReactionRequest) error {
+func RemoveReaction(request DiscussionReactionRequest) error {
 	resp, err := client.Request().SetBody(request).Delete("/discussion/reaction")
 	if err != nil {
 		return err
