--- client/discussion/model/comment_scope_enum.go
@@ -1,8 +1,0 @@
-package model
-
-type CommentScopeEnum int
-
-const (
-	CommentScopeGlobal CommentScopeEnum = iota
-	CommentScopeLocal
-)
