--- client/discussion/model/delete_discussion_comment_request.go
@@ -1,6 +1,0 @@
-package model
-
-type DeleteDiscussionCommentRequest struct {
-	PosterId            int64 `json:"posterId"`
-	DiscussionCommentId int64 `json:"discussionCommentId"`
-}
