--- client/discussion/model/discussion_available_request.go
@@ -1,6 +1,0 @@
-package model
-
-type DiscussionAvailableRequest struct {
-	RepoId    int64 `json:"repoId"`
-	Available bool  `json:"available"`
-}
