--- client/discussion/model/discussion_count_respones.go
@@ -1,6 +1,0 @@
-package model
-
-type DiscussionCountResponse struct {
-	OpenCount  int `json:"openCount"`
-	CloseCount int `json:"closeCount"`
-}
