--- client/discussion/model/discussion_reaction_request.go
@@ -1,8 +1,0 @@
-package model
-
-type DiscussionReactionRequest struct {
-	Type         ReactionTypeEnum `json:"type"`
-	DiscussionId int64            `json:"discussionId"`
-	CommentId    int64            `json:"commentId"`
-	UserId       int64            `json:"userId"`
-}
