--- client/discussion/model/extracted_line.go
@@ -1,6 +1,0 @@
-package model
-
-type ExtractedLine struct {
-	LineNumber int    `json:"lineNumber"`
-	Content    string `json:"content"`
-}
