--- client/discussion/model/file_content.go
@@ -1,6 +1,0 @@
-package model
-
-type FileContent struct {
-	FilePath   string      `json:"filePath"`
-	CodeBlocks []CodeBlock `json:"codeBlocks"`
-}
