--- client/discussion/model/modify_discussion_request.go
@@ -1,10 +1,0 @@
-package model
-
-type ModifyDiscussionRequest struct {
-	RepoId       int64            `json:"repoId"`
-	DiscussionId int64            `json:"discussionId"`
-	PosterId     int64            `json:"posterId"`
-	Name         string           `json:"name"`
-	Content      string           `json:"content"`
-	Codes        []DiscussionCode `json:"codes"`
-}
