--- client/discussion/model/reaction_type_enum.go
@@ -1,14 +1,0 @@
-package model
-
-type ReactionTypeEnum = string
-
-const (
-	PLUS_ONE  ReactionTypeEnum = "+1"
-	MINUS_ONE ReactionTypeEnum = "-1"
-	LAUGH     ReactionTypeEnum = "laugh"
-	HOORAY    ReactionTypeEnum = "hooray"
-	CONFUSED  ReactionTypeEnum = "confused"
-	HEART     ReactionTypeEnum = "heart"
-	ROCKET    ReactionTypeEnum = "rocket"
-	EYES      ReactionTypeEnum = "eyes"
-)
