--- client/discussion/model/update_assignee_request.go
@@ -1,6 +1,0 @@
-package model
-
-type UpdateAssigneeRequest struct {
-	DiscussionId int64 `json:"discussionId"`
-	AssigneeId   int64 `json:"assigneeId"`
-}
