+++ modules/webhook/type.go
@@ -32,7 +32,6 @@ const (
 	HookEventRelease                   HookEventType = "release"
 	HookEventPackage                   HookEventType = "package"
 	HookEventSchedule                  HookEventType = "schedule"
-	HookEventDiscussions               HookEventType = "discussions"
 )
 
 // Event returns the HookEventType as an event string
