+++ routers/web/repo/discussion.go
@@ -16,7 +16,6 @@ import (
 	api "code.gitea.io/gitea/modules/structs"
 
 	discussion_client "code.gitea.io/gitea/client/discussion"
-	"code.gitea.io/gitea/client/discussion/model"
 	issues_model "code.gitea.io/gitea/models/issues"
 	"code.gitea.io/gitea/models/organization"
 	access_model "code.gitea.io/gitea/models/perm/access"
@@ -34,7 +33,6 @@ import (
 	"code.gitea.io/gitea/services/context"
 	discussion_service "code.gitea.io/gitea/services/discussion"
 	"code.gitea.io/gitea/services/forms"
-	notify_service "code.gitea.io/gitea/services/notify"
 )
 
 const (
@@ -64,7 +62,7 @@ func NewDiscussionPost(ctx *context.Context) {
 	if ctx.Written() {
 		return
 	}
-	req := &model.PostDiscussionRequest{
+	req := &discussion_client.PostDiscussionRequest{
 		RepoId:     repo.ID,
 		Poster:     ctx.Doer,
 		PosterId:   ctx.Doer.ID,
@@ -74,11 +72,11 @@ func NewDiscussionPost(ctx *context.Context) {
 		Codes:      form.Codes,
 	}
 	discussionId, err := discussion_service.NewDiscussion(ctx, repo, req)
+	log.Info("New Discussion Post : %v", discussionId)
 	if err != nil {
 		ctx.ServerError("NewDiscussion", err)
 		return
 	}
-	notify_service.NewDiscussion(ctx, ctx.Doer, repo, discussionId)
 	ctx.JSON(http.StatusOK, map[string]interface{}{
 		"discussionId": discussionId,
 	})
@@ -218,18 +216,18 @@ func NewDiscussionCommentPost(ctx *context.Context) {
 	discussionId := ctx.ParamsInt64(":discussionId")
 	commentScope := util.Iif(
 		form.CodeId != nil && form.StartLine != nil && form.EndLine != nil,
-		model.CommentScopeLocal,
-		model.CommentScopeGlobal,
+		discussion_client.CommentScopeLocal,
+		discussion_client.CommentScopeGlobal,
 	)
 	// if request malformed, then make it valid
-	if commentScope == model.CommentScopeGlobal {
+	if commentScope == discussion_client.CommentScopeGlobal {
 		form.CodeId = nil
 		form.StartLine = nil
 		form.EndLine = nil
 		form.GroupId = nil
 	}
 
-	req := &model.PostCommentRequest{
+	req := &discussion_client.PostCommentRequest{
 		DiscussionId: discussionId,
 		Scope:        commentScope,
 		PosterId:     ctx.Doer.ID,
@@ -261,7 +259,7 @@ type DiscussionComment struct {
 	StartLine       int64
 	CodeId          int64
 	EndLine         int64
-	Reactions       model.ReactionList
+	Reactions       discussion_client.ReactionList
 	RenderedContent template.HTML
 	CreatedUnix     timeutil.TimeStamp
 }
@@ -331,7 +329,7 @@ type DiscussionFileCommentsResponse struct {
 
 func groupDiscussionFileCommentsByGroupId(
 	ctx *context.Context,
-	commentsResp []*model.DiscussionCommentResponse) (map[int64][]*DiscussionComment, error) {
+	commentsResp []*discussion_client.DiscussionCommentResponse) (map[int64][]*DiscussionComment, error) {
 
 	comments := make(map[int64][]*DiscussionComment)
 
@@ -516,7 +514,7 @@ func SetDiscussionDeadline(ctx *context.Context) {
 		return
 	}
 
-	ctx.JSON(http.StatusCreated, model.DiscussionDeadline{Deadline: &deadline})
+	ctx.JSON(http.StatusCreated, discussion_client.DiscussionDeadline{Deadline: &deadline})
 }
 
 func getActionDiscussionIds(ctx *context.Context) []int64 {
@@ -581,15 +579,15 @@ func ModifyDiscussionFileComment(ctx *context.Context) {
 
 	commentScope := util.Iif(
 		form.CodeId != nil && form.StartLine != nil && form.EndLine != nil,
-		model.CommentScopeLocal,
-		model.CommentScopeGlobal,
+		discussion_client.CommentScopeLocal,
+		discussion_client.CommentScopeGlobal,
 	)
 
 	posterId := ctx.Doer.ID
 
 	discussionId := ctx.ParamsInt64(":discussionId")
 
-	request := &model.ModifyDiscussionCommentRequest{
+	request := &discussion_client.ModifyDiscussionCommentRequest{
 		DiscussionId:        discussionId,             // discussionId 변수는 이미 전달된 것으로 가정
 		DiscussionCommentId: form.DiscussionCommentId, // form에서 넘어온 DiscussionCommentId
 		CodeId:              form.CodeId,              // form에서 넘어온 CodeId (포인터)
@@ -646,7 +644,7 @@ func ChangeDiscussionCommentReaction(ctx *context.Context) {
 	}
 	action := ctx.Params(":action")
 
-	req := model.DiscussionReactionRequest{
+	req := discussion_client.DiscussionReactionRequest{
 		Type:         form.Content,
 		DiscussionId: discussionId,
 		CommentId:    commentId,
@@ -674,7 +672,7 @@ func ChangeDiscussionCommentReaction(ctx *context.Context) {
 	if err != nil {
 		ctx.ServerError("Failed to Get Discussion Content", err)
 	}
-	var reactions model.ReactionList
+	var reactions discussion_client.ReactionList
 	for _, gc := range d.GlobalComments {
 		if gc.Id == commentId {
 			reactions = gc.Reactions
@@ -695,7 +693,7 @@ func ChangeDiscussionCommentReaction(ctx *context.Context) {
 	})
 }
 
-func discussionRoleDescriptor(ctx stdCtx.Context, repo *repo_model.Repository, poster *user_model.User, discussionResponse *model.DiscussionResponse) (issues_model.RoleDescriptor, error) {
+func discussionRoleDescriptor(ctx stdCtx.Context, repo *repo_model.Repository, poster *user_model.User, discussionResponse *discussion_client.DiscussionResponse) (issues_model.RoleDescriptor, error) {
 	roleDescriptor := issues_model.RoleDescriptor{}
 	roleDescriptor.IsPoster = discussionResponse.IsPoster(poster.ID)
 
@@ -792,7 +790,7 @@ func CreateAiSampleCodeForDiscussion(ctx *context.Context) {
 	ctx.HTML(http.StatusOK, tplDiscussionFileComment)
 }
 
-func convertAiSampleCodeToDiscussionComment(ctx *context.Context, sampleCode *discussion.AiSampleCode, comment *model.DiscussionCommentResponse) (*DiscussionComment, error) {
+func convertAiSampleCodeToDiscussionComment(ctx *context.Context, sampleCode *discussion.AiSampleCode, comment *discussion_client.DiscussionCommentResponse) (*DiscussionComment, error) {
 
 	poster, err := user_model.GetPossibleUserByID(ctx, -3)
 
@@ -828,7 +826,7 @@ func UpdateDiscussionAssignee(ctx *context.Context) {
 			ctx.ServerError("error on discussion response: err = %v", err)
 		}
 	default:
-		req := &model.UpdateAssigneeRequest{
+		req := &discussion_client.UpdateAssigneeRequest{
 			DiscussionId: discussionId,
 			AssigneeId:   assigneeId,
 		}
