+++ services/forms/repo_discussion_form.go
@@ -3,7 +3,7 @@ package forms
 import (
 	"net/http"
 
-	"code.gitea.io/gitea/client/discussion/model"
+	discussion_client "code.gitea.io/gitea/client/discussion"
 	"code.gitea.io/gitea/modules/web/middleware"
 	"code.gitea.io/gitea/services/context"
 	"gitea.com/go-chi/binding"
@@ -13,7 +13,7 @@ type CreateDiscussionForm struct {
 	Name       string
 	Content    string
 	BranchName string
-	Codes      []model.DiscussionCode
+	Codes      []discussion_client.DiscussionCode
 }
 
 func (d *CreateDiscussionForm) Validate(req *http.Request, errs binding.Errors) binding.Errors {
