+++ templates/user/notification/notification_div.tmpl
@@ -35,7 +35,6 @@
 					</div>
 				{{else}}
 					{{range $notification := .Notifications}}
-						{{if ne .IssueID 0}}
 						<div class="notifications-item tw-flex tw-items-center tw-flex-wrap tw-gap-2 tw-p-2" id="notification_{{.ID}}" data-status="{{.Status}}">
 							<div class="notifications-icon tw-ml-2 tw-mr-1 tw-self-start tw-mt-1">
 								{{if .Issue}}
@@ -117,37 +116,6 @@
 								{{end}}
 							</div>
 						</div>
-						{{else if ne .DiscussionID 0}}
-						<div class="notifications-item tw-flex tw-items-center tw-flex-wrap tw-gap-2 tw-p-2" id="notification_{{.ID}}" data-status="{{.Status}}">
-							<div class="notifications-icon tw-ml-2 tw-mr-1 tw-self-start tw-mt-1">
-								{{svg "octicon-comment-discussion" 16 "text grey"}}
-							</div>
-							<a class="notifications-link tw-flex tw-flex-1 tw-flex-col silenced" href="/{{.Repository.FullName}}/discussions/{{.Discussion.Id}}">
-								<div class="notifications-top-row tw-text-13">
-									{{.Repository.FullName}}
-									<span class="text light-3">#{{.Discussion.Index}}</span>
-									{{if eq .Status 3}}
-										{{svg "octicon-pin" 13 "text blue tw-mt-0.5 tw-ml-1"}}
-									{{end}}
-								</div>
-								<div class="notifications-bottom-row tw-text-15 tw-py-0.5">
-									<span class="issue-title">
-										{{.Discussion.Name | RenderEmoji $.Context | RenderCodeBlock}}
-									</span>
-								</div>
-							</a>
-							<div class="notifications-updated tw-items-center tw-mr-2">
-								{{TimeSinceUnix .Discussion.UpdatedUnix ctx.Locale}}
-							</div>
-							
-							<!-- 
-							TODO: handle mouse over
-							<div class="notifications-buttons tw-items-center tw-justify-end tw-gap-1 tw-px-1">
-							
-							</div> 
-							-->
-						</div>
-						{{end}}
 					{{end}}
 				{{end}}
 			</div>
